recherche produit et promoting par categorie 1.0

+++ src/main/java/com/example/springjee/entities/Produit.java
@@ -12,7 +12,7 @@ public class Produit {
     @Id
     @Column(name = "produit_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "produit_reference", nullable = false)
     private String reference;
@@ -62,11 +62,9 @@ public class Produit {
     //////GETTER AND SETTER//////
     /////////////////////////////
 
-    public Long getId() {
-        return id;
-    }
+    public Integer getId() { return id; }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
